+++ core/common/src/main/java/dev/enesky/core/common/utils/Constants.kt
@@ -22,4 +22,6 @@ package dev.enesky.core.common.utils
 object Constants {
 
     const val ITEMS_PER_PAGE = 20
+    const val PLACEHOLDER_IMG_URL = "https://fakeimg.pl/400x600/2b323f/FFFFFF?text=Not+found"
+
 }
