+++ core/design-system/src/main/java/dev/enesky/core/design_system/home/AnimeItem.kt
@@ -17,14 +17,13 @@ import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import dev.enesky.core.data.models.ImageList
-import dev.enesky.core.data.models.Images
 import dev.enesky.core.design_system.common.DoodleImagePlaceholder
 import dev.enesky.core.design_system.common.DoodleNetworkImage
 import dev.enesky.core.design_system.common.placeholder
 import dev.enesky.core.design_system.theme.DoodleTheme
 import dev.enesky.core.design_system.theme.Icons
 import dev.enesky.core.domain.models.Anime
+import dev.enesky.core.domain.models.placeholderAnime
 
 /**
  * Created by Enes Kamil YILMAZ on 21/12/2023
@@ -65,7 +64,7 @@ fun AnimeItem(
                 modifier = Modifier
                     .size(itemWidth, itemHeight)
                     .clip(DoodleTheme.shapes.small),
-                model = anime.images.jpg?.imageUrl,
+                model = anime.imageUrl,
                 contentDescription = anime.title,
             )
         }
@@ -125,18 +124,7 @@ fun AnimeItem(
 private fun AnimeItemPreview() {
     DoodleTheme {
         AnimeItem(
-            anime = Anime(
-                id = 0,
-                title = "Jujutsu Kaisen",
-                genres = "Action | Adventure",
-                trailer = null,
-                url = "",
-                images = Images(
-                    jpg = ImageList(
-                        imageUrl = "",
-                    ),
-                ),
-            ),
+            anime = placeholderAnime,
             isPlaceholder = false,
         )
     }
