+++ core/design-system/src/main/java/dev/enesky/core/design_system/home/TopAnimePreview.kt
@@ -18,13 +18,12 @@ import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import dev.enesky.core.data.models.ImageList
-import dev.enesky.core.data.models.Images
 import dev.enesky.core.design_system.R
 import dev.enesky.core.design_system.common.DoodleImagePlaceholder
 import dev.enesky.core.design_system.common.DoodleNetworkImage
 import dev.enesky.core.design_system.theme.DoodleTheme
 import dev.enesky.core.domain.models.Anime
+import dev.enesky.core.domain.models.placeholderAnime
 
 /**
  * Created by Enes Kamil YILMAZ on 20/12/2023
@@ -57,7 +56,7 @@ fun TopAnimePreview(
                 }
             },
     ) {
-        if (isLoading || anime == null || anime?.images?.jpg?.imageUrl == null) {
+        if (isLoading || anime == null) {
             DoodleImagePlaceholder(
                 modifier = Modifier
                     .height(itemHeight)
@@ -74,7 +73,7 @@ fun TopAnimePreview(
             DoodleNetworkImage(
                 modifier = Modifier
                     .height(itemHeight),
-                model = anime.images.jpg?.imageUrl,
+                model = anime.imageUrl,
                 contentDescription = anime.title,
             )
         }
@@ -122,18 +121,7 @@ fun TopAnimePreview(
 private fun TopAnimePreviewPreview() {
     DoodleTheme {
         TopAnimePreview(
-            anime = Anime(
-                id = 0,
-                title = "Jujutsu Kaisen",
-                genres = "Action | Adventure",
-                trailer = null,
-                url = "",
-                images = Images(
-                    jpg = ImageList(
-                        imageUrl = "",
-                    ),
-                ),
-            ),
+            anime = placeholderAnime,
         ) {}
     }
 }
