+++ core/domain/src/main/java/dev/enesky/core/domain/mappers/AnimeMapper.kt
@@ -3,7 +3,6 @@ package dev.enesky.core.domain.mappers
 import dev.enesky.core.data.response.AnimeResponse
 import dev.enesky.core.data.response.FullAnime
 import dev.enesky.core.domain.models.Anime
-import dev.enesky.core.domain.toGenreString
 
 /**
  * Created by Enes Kamil YILMAZ on 23/12/2023
@@ -13,7 +12,7 @@ fun AnimeResponse.asAnime(): Anime {
     return Anime(
         id = id,
         url = url,
-        images = images,
+        imageUrl = getImageUrl(images),
         trailer = trailer,
         title = title,
         genres = toGenreString(genres),
@@ -23,7 +22,7 @@ fun AnimeResponse.asAnime(): Anime {
 fun FullAnime.asAnime() = Anime(
     id = id,
     url = url,
-    images = images,
+    imageUrl = getImageUrl(images),
     trailer = trailer,
     title = title,
     genres = toGenreString(genres),
