+++ core/domain/src/main/java/dev/enesky/core/domain/models/Anime.kt
@@ -2,7 +2,6 @@ package dev.enesky.core.domain.models
 
 import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
-import dev.enesky.core.data.models.Images
 import dev.enesky.core.data.response.Trailer
 import kotlinx.parcelize.Parcelize
 
@@ -14,8 +13,19 @@ import kotlinx.parcelize.Parcelize
 data class Anime(
     @SerializedName("mal_id") val id: Int,
     val url: String,
-    val images: Images,
     val trailer: Trailer? = null,
     val title: String,
     val genres: String,
+    val imageUrl: String
 ) : Parcelable
+
+// Placeholder Anime for preview
+val placeholderAnime: Anime
+    get() = Anime(
+        id = 0,
+        title = "Jujutsu Kaisen",
+        genres = "Action | Adventure",
+        trailer = null,
+        url = "",
+        imageUrl = "",
+    )
